+++ docs/release_notes.rst
@@ -1,6 +1,18 @@
 Release Notes
 ---------------
 
+v1.0.1 April 7, 2024
+====================
+
+- **Enhacements:**
+    - Set default tournament to 10 for GeneticFeatureSynthesis class
+    - Added `tournament_size` parameter to GeneticFeatureSelection class and set default to 10
+- **Documentation:**
+    - Updated README.md example
+    - Updated example notebooks
+    - Added eplanation of tournament selection in the `Tuning the Genetic Feature Synthesis` guide
+
+
 v1.0.0 Mar 30, 2024
 ====================